+++ utils/check_repo.py
@@ -51,6 +51,7 @@ PRIVATE_MODELS = [
     "MaskFormerSwinPreTrainedModel",
     "BridgeTowerTextModel",
     "BridgeTowerVisionModel",
+    "BlipTextLMHeadModel",
 ]
 
 # Update this list for models that are not tested with a comment explaining the reason it should not be.
